pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "jsonpatchplus"
description = "JsonPatch+"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyYAML",
]
license = {file = "LICENSE"}
authors = [
    {name = "Elmer Nocon", email = "elmernocon@gmail.com"},
]
dynamic = [
    "version"
]

[project.optional-dependencies]
dev = [
    "black",
    "build",
    "twine",
]
test = [
    "deepdiff",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "jsonpatchplus.__version__"}
